Upgraded shared preferences flow
