+++ android/app/src/main/java/updated/mysterium/vpn/network/usecase/BalanceUseCase.kt
@@ -1,16 +1,17 @@
 package updated.mysterium.vpn.network.usecase
 
-import android.content.Context
 import mysterium.GetBalanceRequest
 import network.mysterium.service.core.DeferredNode
 import network.mysterium.service.core.NodeRepository
+import updated.mysterium.vpn.database.preferences.SharedPreferencesManager
 
-class BalanceUseCase(private val nodeRepository: NodeRepository, private val context: Context) {
+class BalanceUseCase(
+    private val nodeRepository: NodeRepository,
+    private val sharedPreferencesManager: SharedPreferencesManager
+) {
 
     private companion object {
         const val CURRENCY = "USD"
-        const val BALANCE_POP_UP_PREFERENCES_KEY = "BALANCE_POP_UP"
-        const val BALANCE_POP_UP_SHOWN_KEY = "BALANCE_POP_UP_SHOWN"
     }
 
     fun initDeferredNode(deferredNode: DeferredNode) {
@@ -27,12 +28,7 @@ class BalanceUseCase(private val nodeRepository: NodeRepository, private val con
 
     suspend fun getUsdEquivalent() = nodeRepository.getExchangeRate(CURRENCY)
 
-    fun isBalancePopUpShown() = context.getSharedPreferences(
-        BALANCE_POP_UP_PREFERENCES_KEY, Context.MODE_PRIVATE
-    ).contains(BALANCE_POP_UP_SHOWN_KEY)
+    fun isBalancePopUpShown() = sharedPreferencesManager.isBalancePopUpShown()
 
-    fun balancePopUpShown() = context
-        .getSharedPreferences(BALANCE_POP_UP_PREFERENCES_KEY, Context.MODE_PRIVATE)
-        .edit()
-        .putBoolean(BALANCE_POP_UP_SHOWN_KEY, true).apply()
+    fun balancePopUpShown() = sharedPreferencesManager.balancePopUpShown()
 }
